requirements.txt
python-telegram-bot>=20.0
aiohttp
python-dotenv
fastapi
uvicorn